refactor: 컴포넌트 내 폴더명 수정
- utils -> tools

[#802]
