+++ frontend/src/components/ui/StationSearchWindow/tools/convertStationDetailsToSummary.ts
